refactor: reorganize admin routes into app/routers/admin/
