+++ backend/tests/routers/test_contact.py
@@ -42,102 +42,6 @@ def test_send_message(client: TestClient, session: Session) -> None:
     assert msg.read is False
 
 
-def test_list_messages_admin(admin_client: TestClient, client: TestClient) -> None:
-    client.post(
-        CONTACT_URL,
-        json={
-            "name": "first",
-            "email": "first@test.com",
-            "subject": "sub1",
-            "message": "msg1",
-        },
-    )
-    client.post(
-        CONTACT_URL,
-        json={
-            "name": "second",
-            "email": "second@test.com",
-            "subject": "sub2",
-            "message": "msg2",
-        },
-    )
-
-    response = admin_client.get(CONTACT_URL)
-    assert response.status_code == 200
-
-    data = response.json()
-    assert len(data) == 2
-    assert data[0]["name"] == "first"
-    assert data[1]["name"] == "second"
-    assert "id" in data[0]
-    assert "read" in data[0]
-    assert "created_at" in data[0]
-
-
-def test_list_messages_without_auth(client: TestClient) -> None:
-    response = client.get(CONTACT_URL)
-    assert response.status_code == 401
-
-
-def test_mark_as_read(
-    session: Session,
-    admin_client: TestClient,
-    client: TestClient,
-    message: ContactMessage,
-) -> None:
-    response = admin_client.patch(f"{CONTACT_URL}/{message.id}/read")
-    assert response.status_code == 200
-    assert message.read is True
-
-
-def test_mark_as_read_not_found(admin_client: TestClient) -> None:
-    response = admin_client.patch(f"{CONTACT_URL}/999/read")
-    assert response.status_code == 404
-
-
-def test_delete_message(admin_client: TestClient, message: ContactMessage) -> None:
-    response = admin_client.delete(f"{CONTACT_URL}/{message.id}")
-    assert response.status_code == 204
-
-
-def test_deleted_message_hidden_from_list(
-    admin_client: TestClient, message: ContactMessage
-) -> None:
-    admin_client.delete(f"{CONTACT_URL}/{message.id}")
-    response = admin_client.get(CONTACT_URL)
-    assert response.status_code == 200
-    assert len(response.json()) == 0
-
-
-def test_delete_message_not_found(admin_client: TestClient) -> None:
-    response = admin_client.delete(f"{CONTACT_URL}/999")
-    assert response.status_code == 404
-
-
-def test_delete_message_without_auth(
-    client: TestClient, message: ContactMessage
-) -> None:
-    response = client.delete(f"{CONTACT_URL}/{message.id}")
-    assert response.status_code == 401
-
-
-def test_list_messages_pagination(admin_client: TestClient, session: Session) -> None:
-    for i in range(5):
-        session.add(
-            ContactMessage(
-                name=f"user{i}",
-                email=f"u{i}@test.com",
-                subject="sub",
-                message="msg",
-            )
-        )
-    session.commit()
-
-    response = admin_client.get(f"{CONTACT_URL}?limit=2")
-    assert response.status_code == 200
-    assert len(response.json()) == 2
-
-
 def test_create_message_rate_limited(client: TestClient) -> None:
     payload = {
         "name": "spam",
@@ -189,22 +93,3 @@ def test_send_message_too_long(client: TestClient) -> None:
         },
     )
     assert response.status_code == 422
-
-
-def test_list_messages_pagination_offset_beyond_total(
-    admin_client: TestClient, session: Session
-) -> None:
-    for i in range(3):
-        session.add(
-            ContactMessage(
-                name=f"user{i}",
-                email=f"u{i}@test.com",
-                subject="sub",
-                message="msg",
-            )
-        )
-        session.commit()
-
-        response = admin_client.get(f"{CONTACT_URL}?offset=10")
-        assert response.status_code == 200
-        assert response.json() == []
